+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * feat(backups-config): Warn user before unscheduling periodic backups
 * feat(collaborators): Update list to include `is_limited` information
+* feat(collaborators): add command to update a collaborator
 
 ### 1.35.1